pyproject.toml
[tool.poetry]
name = "cslistipy"
version = "0.1.0"
description = ""
authors = ["Andreas Løfsgaard <andreas@lofsgaard.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
spotipy = "^2.22.0"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

